♻️ refactor: 코드 리팩토링+ 폴더 이동 + api 연동
