+++ app/(route)/admin/@modal/(.)notice-create/_components/schema.ts
